Dynamically construct file listing
Instead of generated the HTML table on the server,
simply embed a JSON listing of all files and
use Javascript to render the table.

This now enables support for sorting the list by column,
but also requires us to now implement the formatting logic
in Javascript rather than in Go.